update: extracted some code blocks to different files
